requirements.txt
fastapi
jsql
python-dotenv
mysql-connector-python
sqlalchemy
uvicorn
jinja2